Cleaner code using presenters
